requirements.txt
python-dotenv==1.0.0
boto3==1.34.0
duckdb==0.10.0
psycopg2-binary==2.9.9
requests==2.31.0
pandas==2.1.4